+++ tests/test_preprocessor.py
@@ -1,11 +1,16 @@
 from io import BytesIO
 from pathlib import Path
 
+import holoviews as hv
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 
-from enjoyn.preprocessor import MatplotlibPreprocessor, Preprocessor
+from enjoyn.preprocessor import (
+    HoloViewsPreprocessor,
+    MatplotlibPreprocessor,
+    Preprocessor,
+)
 
 
 class TestPreprocessor:
@@ -25,14 +30,15 @@ class TestPreprocessor:
     def test_apply_on(self, return_obj):
         func = lambda item: return_obj
         preprocessor = Preprocessor(func=func)
-        actual = preprocessor.apply_on("irrelevant_item")
-        assert actual == return_obj
+        with preprocessor.apply_on("irrelevant_item") as actual:
+            assert actual == return_obj
 
     def test_apply_on_type_error(self):
         func = lambda item: 1234
         preprocessor = Preprocessor(func=func)
         with pytest.raises(TypeError, match="callable returned an object with"):
-            preprocessor.apply_on("irrelevant_item")
+            with preprocessor.apply_on("irrelevant_item"):
+                pass
 
 
 class TestMatplotlibPreprocessor:
@@ -45,8 +51,9 @@ class TestMatplotlibPreprocessor:
             return (fig, ax, img)[return_index]
 
         preprocessor = MatplotlibPreprocessor(func=plot)
-        actual = preprocessor.apply_on(10)
-        assert isinstance(actual, BytesIO)
+        with preprocessor.apply_on(10) as actual:
+            assert isinstance(actual, BytesIO)
+        assert actual.closed
 
     def test_apply_on_plt(self):
         def plot(i):
@@ -54,11 +61,29 @@ class TestMatplotlibPreprocessor:
             return img
 
         preprocessor = MatplotlibPreprocessor(func=plot)
-        actual = preprocessor.apply_on(10)
-        assert isinstance(actual, BytesIO)
+        with preprocessor.apply_on(10) as actual:
+            assert isinstance(actual, BytesIO)
+        assert actual.closed
 
     def test_apply_not_matplotlib(self):
         func = lambda item: Path("path")
         preprocessor = MatplotlibPreprocessor(func=func)
-        actual = preprocessor.apply_on("irrelevant_item")
-        assert isinstance(actual, Path)
+        with preprocessor.apply_on("irrelevant_item") as actual:
+            assert isinstance(actual, Path)
+
+
+class TestHoloViewsPreprocessor:
+    def test_apply_on(self):
+        def plot(i):
+            return hv.Curve(([0, 1, i], [0, 1, i]))
+
+        preprocessor = HoloViewsPreprocessor(func=plot)
+        with preprocessor.apply_on(10) as actual:
+            assert isinstance(actual, BytesIO)
+        assert actual.closed
+
+    def test_apply_not_holoviews(self):
+        func = lambda item: Path("path")
+        preprocessor = HoloViewsPreprocessor(func=func)
+        with preprocessor.apply_on("irrelevant_item") as actual:
+            assert isinstance(actual, Path)